7th - KeyInput fix, SmartEnemy

+++ src/com/game/BasicEnemy.java
@@ -9,14 +9,14 @@ public class BasicEnemy extends GameObject {
     public BasicEnemy(int x, int y, ID id, Handler handler) {
         super(x, y, id);
 
-        this.handler =handler;
+        this.handler = handler;
 
         velX = 4;
         velY = 4;
     }
 
     public Rectangle getBounds() {
-        return new Rectangle(x, y, 16, 16);
+        return new Rectangle((int) x, (int) y, 16, 16);
     }
 
     public void tick() {
@@ -26,11 +26,11 @@ public class BasicEnemy extends GameObject {
         if (x <= 0 || x >= Game.WIDTH - 16) velX *= -1;
         if (y <= 0 || y >= Game.HEIGHT - 16) velY *= -1;
 
-        handler.addObject(new Trail(x,y,ID.Trail,handler,Color.red,16,16, 0.05f));
+        handler.addObject(new Trail(x, y, ID.Trail, handler, Color.red, 16, 16, 0.05f));
     }
 
     public void render(Graphics g) {
         g.setColor(Color.RED);
-        g.fillRect(x, y, 16, 16);
+        g.fillRect((int) x, (int) y, 16, 16);
     }
 }
